Add meta for system management tool(s).
One of the most frustrating problems in system administration is the
lack of an easy way to save a "known good" state. Tools like etckeeper
and saltstack make half-assed attempts at this, but they're both
cumbersome and require infrastructure to already be in place. This is
an attempt at defining a "better" paradigm, a tool that is clean,
simple, and works at the shell level, without extras or dependencies.