in the beginning, there was a hacker
See, hacking gets a bad name these days, with the media conflating
"hacker" and "cracker".

See, a hacker looks at something, and thinks "hmm, that's odd", or
"...I bet I could get this to do something _else_, too.".

A cracker uses that same mentality for the purpose of causing harm.
E.g., a cracker is a malicious hacker.
There's debate about defs.

I'lll be using that one in this document.
Open pull request if you disagree, and improve something in the
process of being right.