Fix unexpected token syntax error.
When running 'shutgrade', the command returns:
^$ shutgrade
bash: syntax error near unexpected token `;;'

This was caused by ending a command alias in a semicolon, and then using the alias with a semicolon, resulting in the actual command having double semicolons, resulting in the aforementioned bash error.

The uprepair command will also, in theory, benefit from this bugfix.